kafka-html-pdf/consumer/requirements.txt
kafka-python==2.0.2
jinja2
weasyprint
